entz-playground: only get decision on change
Now, rather than polling OPA for a decision every 5 seconds, the
back-end keeps a counter of how many times the bundle has updated, and
exposes this as an API endpoint. The frontend polls this every 500ms,
and retrieves all decisions again if it has changed. The frontend also
re-retrieves all decisions any time a new watcher is added.

STY-10629