redoc: temporary workaround
It seems like redoc-cli serve is no longer working for reasons I cannot
determine. Also, redoc-cli bundle apparently only works if it's output
is not re-directed to a file (???).

As a temporary workaround, I am having redoc build a static HTML file
and serving it with Python's built-in HTTP server. This should probably
be cleaned up a bit in future.

STY-0